requirements.txt
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
inexactsearch==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
outcome==1.2.0
pyspellchecker==0.6.3
python-dotenv==0.20.0
requests==2.28.1
selenium==4.4.3
silpa-common==0.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.3.2.post1
spellchecker==0.4
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
